backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "asyncpg>=0.29.0",
    "fastapi[standard]>=0.121.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
